workshops/MAX-Model-Wrapping/solution/requirements.txt
tensorflow==1.15.2
numpy==1.14.1
Pillow==5.4.1
h5py==2.9.0